Update README.md
I have updated the README.md with essential documentation as a user who has recently started exploring this repository and its features.

- Troubleshooting
I added this section as I encountered problems setting up the localhost server for this repository. I found that although I had the dependencies stated a lot of them were out of date and some were not mentioned. The following errors I found are listed in this section along with the commands I used to fix them after further research. I believe this will be valuable to other users who may encounter similar problems.

- Contributing
I added a section about making a good pull request because I am new to contributing to open source projects and looked up what made a good README.md in as I was adding to the troubleshooting section. I made a summary of the things I learned from other large open source projects. I believe this will aid in the goals of this repository as it keeps with the current contribution standards and adds an extra layer of organization to future pull requests for the ease of developers and future contributors. 

-Usages
I added this new section because when I first visited this page I had a lot of questions about what this repository actually did. I felt that the current wording was somewhat vague in its description and was only able to find out more by setting up the repository. I believe that by adding these screenshots of the different features available users will be more likely to be intrigued by this project and willing to help contribute to it. Furthermore, users will be more likely to make use of this project because they will know exactly what they are signing up for.

I hope this helps the future trajectory of this project. Please let me know if you have any questions!